requirements.txt
Django==1.7.1
djangorestframework==2.4.4
dj-static==0.0.6
dj-database-url==0.3.0
django-bower==5.0.1
